fix issue with responses incorrectly showing as human graded
