+++ services/QuillLMS/client/app/bundles/Connect/components/questions/questions.jsx
@@ -6,12 +6,12 @@ import _ from 'underscore';
 import {
   ArchivedButton,
   Modal,
-  hashToCollection
+  hashToCollection,
+  responsesWithStatus
 } from '../../../Shared/index';
 import actions from '../../actions/questions';
 import { deleteResponse, submitResponseEdit } from '../../actions/responses';
 import Question from '../../libs/question';
-import respWithStatus from '../../libs/responseTools.js';
 import { QuestionListByConcept } from '../shared/questionListByConcept';
 
 function sleep(milliseconds) {
@@ -137,7 +137,7 @@ class Questions extends React.Component {
 
   responsesWithStatusForQuestion = questionUID => {
     const responses = this.props.responses.data[questionUID];
-    return hashToCollection(respWithStatus(responses));
+    return hashToCollection(responsesWithStatus(responses));
   };
 
   submitNewQuestion = () => {
