+++ services/QuillLMS/client/app/bundles/Diagnostic/components/misc/answerVisualizer.jsx
@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import _ from 'underscore';
 import { listenToResponsesWithCallback } from '../../actions/responses';
-import respWithStatus from '../../libs/responseTools.js';
+import { responsesWithStatus, } from '../../../Shared/index';
 import DiffedResponse from './diffedResponse.jsx';
 
 export class AnswerVisualizer extends Component {
@@ -22,7 +22,7 @@ export class AnswerVisualizer extends Component {
       questionID,
       (data) => {
         this.setState({
-          responses: respWithStatus(data),
+          responses: responsesWithStatus(data),
           loadedResponses: true,
         });
       }
