+++ services/QuillLMS/client/app/bundles/Diagnostic/components/questions/responseComponent.jsx
@@ -6,7 +6,8 @@ import {
   QuestionBar,
   ResponseSortFields,
   ResponseToggleFields,
-  hashToCollection
+  hashToCollection,
+  responsesWithStatus,
 } from '../../../Shared/index';
 import * as filterActions from '../../actions/filters';
 import massEdit from '../../actions/massEdit';
@@ -19,7 +20,6 @@ import {
 } from '../../libs/grading/rematching.ts';
 import { getPartsOfSpeechTags } from '../../libs/partsOfSpeechTagging.js';
 import QuestionMatcher from '../../libs/question';
-import respWithStatus from '../../libs/responseTools.js';
 import POSMatcher from '../../libs/sentenceFragment.js';
 import POSForResponsesList from './POSForResponsesList.jsx';
 import ResponseList from './responseList.jsx';
@@ -178,7 +178,7 @@ class ResponseComponent extends React.Component {
   };
 
   responsesWithStatus = () => {
-    return hashToCollection(respWithStatus(this.props.filters.responses));
+    return hashToCollection(responsesWithStatus(this.props.filters.responses));
   };
 
   responsesGroupedByStatus = () => {
