+++ services/QuillLMS/client/app/bundles/Shared/index.tsx
@@ -115,7 +115,10 @@ export {
   filterNumbers,
   redirectToActivity,
   renderNavList,
-  noResultsMessage
+  noResultsMessage,
+  getStatusForResponse,
+  responsesWithStatus,
+  sortByLevenshteinAndOptimal,
 } from './libs/index'
 
 export {
